(chore) Test image filters deletion by an authenticated user

+++ editor/views.py
@@ -96,6 +96,8 @@ def remove_effects(request):
         temp_url = 'static/media/temp/'
         file_list = os.listdir(temp_url)
         for file_name in file_list:
+            if file_name == ".gitignore":
+                continue
             os.remove(temp_url + "/" + file_name)
 
         return Response(status=status.HTTP_200_OK)
